Adding automation for Internal Packages, Dynamo Text
- Moved PackagesValidationTests under WorkflowTests project so that we can have Geometry validation.
- Wrote new test case for testing Text node from Dynamo Text package.
- Addressing this task http://adsk-oss.myjetbrains.com/youtrack/issue/MAGN-7360
