+++ test/DynamoCoreTests/DynamoCoreTests.csproj
@@ -80,7 +80,6 @@
     <Compile Include="NodeToCodeTest.cs" />
     <Compile Include="PackageManager\PackageTests.cs" />
     <Compile Include="PackageManager\RecordedFileSystem.cs" />
-    <Compile Include="PackageValidationTest.cs" />
     <Compile Include="SearchMemberGroupTests.cs" />
     <Compile Include="TestZeroTouchClass.cs" />
     <Compile Include="TypedParametersToStringTests.cs" />
